+++ pkg/vm/engine/tae/compute/compare.go
@@ -22,7 +22,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/container/types"
 )
 
-func CompareOrdered[T types.OrderedT](a, b T) int64 {
+func CompareOrdered[T types.OrderedT](a, b T) int {
 	if a > b {
 		return 1
 	} else if a < b {
@@ -31,7 +31,7 @@ func CompareOrdered[T types.OrderedT](a, b T) int64 {
 	return 0
 }
 
-func CompareBool(a, b bool) int64 {
+func CompareBool(a, b bool) int {
 	if a == b {
 		return 0
 	} else if a {
@@ -40,7 +40,7 @@ func CompareBool(a, b bool) int64 {
 	return -1
 }
 
-func CompareBytes(a, b []byte) int64 {
+func CompareBytes(a, b []byte) int {
 	res := bytes.Compare(a, b)
 	if res > 0 {
 		return 1
@@ -51,7 +51,7 @@ func CompareBytes(a, b []byte) int64 {
 	}
 }
 
-func Compare(a, b []byte, t types.T, scale1, scale2 int32) int64 {
+func Compare(a, b []byte, t types.T, scale1, scale2 int32) int {
 	switch t {
 	case types.T_bool:
 		return CompareBool(types.DecodeBool(a), types.DecodeBool(b))
@@ -105,7 +105,7 @@ func Compare(a, b []byte, t types.T, scale1, scale2 int32) int64 {
 	}
 }
 
-func CompareGeneric(a, b any, t types.T) int64 {
+func CompareGeneric(a, b any, t types.T) int {
 	switch t {
 	case types.T_bool:
 		return CompareBool(a.(bool), b.(bool))
@@ -126,11 +126,11 @@ func CompareGeneric(a, b any, t types.T) int64 {
 	case types.T_uint64:
 		return CompareOrdered(a.(uint64), b.(uint64))
 	case types.T_decimal64:
-		return int64(a.(types.Decimal64).Compare(b.(types.Decimal64)))
+		return a.(types.Decimal64).Compare(b.(types.Decimal64))
 	case types.T_decimal128:
-		return int64(a.(types.Decimal128).Compare(b.(types.Decimal128)))
+		return a.(types.Decimal128).Compare(b.(types.Decimal128))
 	case types.T_decimal256:
-		return int64(a.(types.Decimal256).Compare(b.(types.Decimal256)))
+		return a.(types.Decimal256).Compare(b.(types.Decimal256))
 	case types.T_float32:
 		return CompareOrdered(a.(float32), b.(float32))
 	case types.T_float64:
@@ -144,7 +144,7 @@ func CompareGeneric(a, b any, t types.T) int64 {
 	case types.T_datetime:
 		return CompareOrdered(a.(types.Datetime), b.(types.Datetime))
 	case types.T_TS:
-		return int64(a.(types.TS).Compare(b.(types.TS)))
+		return a.(types.TS).Compare(b.(types.TS))
 	case types.T_Rowid:
 		return CompareBytes(a.([]byte), b.([]byte))
 	case types.T_Blockid:
